requirements.txt
boto3==1.4.4
botocore==1.5.31
coreapi==2.3.0
coreschema==0.0.4
Django==1.10.6
djangorestframework==3.6.2
docutils==0.13.1
isodate==0.5.4
itypes==1.1.0
Jinja2==2.9.5
jmespath==0.9.2
Markdown==2.6.8
MarkupSafe==1.0
numpy==1.12.1
olefile==0.44
opencv-python==3.2.0.7
pandas==0.19.2
Pillow==4.1.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
rdflib==4.2.2
requests==2.13.0
s3transfer==0.1.10
six==1.10.0
SPARQLWrapper==1.8.0
tqdm==4.11.2
uritemplate==3.0.0
